Add an option to use json file as credential source
A number of k8s secrets providers will expose an encrypted volume mount with a
file path with which to provide secrets. This allows a file path to be
defined as IB_CREDENTIALS_FILEPATH with a JSON dict containing "account"
"password" and "trade_mode" as keys for their respective params.

I also eliminated the GCP_SECRET env var usage since it seemed redundant
when GOOGLE_APPLICATION_CREDENTIALS is defined.

I had trouble running the tests locally, as I am using podman and the
docker socket did not seem to have the right permissions.

I do not have access to a GCP setup so help testing for regressions
would be appreciated